+++ src/aux-common/Files/FilesChannel.spec.ts
@@ -10,6 +10,7 @@ import {
     FileAddedEvent,
     toast,
     tweenTo,
+    openQRCodeScanner,
 } from './FilesChannel';
 import { File } from './File';
 import uuid from 'uuid/v4';
@@ -1516,7 +1517,7 @@ describe('FilesChannel', () => {
                         id: 'thisFile',
                         tags: {
                             'addItem()':
-                                'applyDiff(@name("bob"), makeDiff.addToMenu())',
+                                'applyDiff(@name("bob"), diff.addToMenu())',
                         },
                     },
                     userFile: {
@@ -1615,7 +1616,7 @@ describe('FilesChannel', () => {
                         id: 'thisFile',
                         tags: {
                             'addItem()':
-                                'applyDiff(@name("bob"), makeDiff.removeFromMenu())',
+                                'applyDiff(@name("bob"), diff.removeFromMenu())',
                         },
                     },
                     userFile: {
@@ -1819,7 +1820,7 @@ describe('FilesChannel', () => {
                         id: 'thisFile',
                         tags: {
                             'test()':
-                                'applyDiff(this, makeDiff.addToContext("abc"))',
+                                'applyDiff(this, diff.addToContext("abc"))',
                         },
                     },
                 };
@@ -1852,7 +1853,7 @@ describe('FilesChannel', () => {
                         tags: {
                             abc: true,
                             'test()':
-                                'applyDiff(this, makeDiff.removeFromContext("abc"))',
+                                'applyDiff(this, diff.removeFromContext("abc"))',
                         },
                     },
                 };
@@ -1948,7 +1949,7 @@ describe('FilesChannel', () => {
                         id: 'thisFile',
                         tags: {
                             'test()':
-                                'applyDiff(this, makeDiff.setPosition("abc", 1, 2))',
+                                'applyDiff(this, diff.setPosition("abc", 1, 2))',
                         },
                     },
                 };
@@ -1976,7 +1977,7 @@ describe('FilesChannel', () => {
                         id: 'thisFile',
                         tags: {
                             'test()':
-                                'applyDiff(this, makeDiff.setPosition("abc", undefined, 2))',
+                                'applyDiff(this, diff.setPosition("abc", undefined, 2))',
                         },
                     },
                 };
@@ -2003,7 +2004,7 @@ describe('FilesChannel', () => {
                         id: 'thisFile',
                         tags: {
                             'test()':
-                                'applyDiff(this, makeDiff.setPosition("abc", undefined, undefined, 2))',
+                                'applyDiff(this, diff.setPosition("abc", undefined, undefined, 2))',
                         },
                     },
                 };
@@ -2124,6 +2125,50 @@ describe('FilesChannel', () => {
                 expect(result.hasUserDefinedEvents).toBe(true);
 
                 expect(result.events).toEqual([tweenTo('thisFile')]);
+            });
+        });
+
+        describe('openQRCodeScanner()', () => {
+            it('should emit a OpenQRCodeScannerEvent', () => {
+                const state: FilesState = {
+                    thisFile: {
+                        id: 'thisFile',
+                        tags: {
+                            'test()': 'openQRCodeScanner()',
+                        },
+                    },
+                };
+
+                // specify the UUID to use next
+                uuidMock.mockReturnValue('uuid-0');
+                const fileAction = action('test', ['thisFile']);
+                const result = calculateActionEvents(state, fileAction);
+
+                expect(result.hasUserDefinedEvents).toBe(true);
+
+                expect(result.events).toEqual([openQRCodeScanner(true)]);
+            });
+        });
+
+        describe('closeQRCodeScanner()', () => {
+            it('should emit a OpenQRCodeScannerEvent', () => {
+                const state: FilesState = {
+                    thisFile: {
+                        id: 'thisFile',
+                        tags: {
+                            'test()': 'closeQRCodeScanner()',
+                        },
+                    },
+                };
+
+                // specify the UUID to use next
+                uuidMock.mockReturnValue('uuid-0');
+                const fileAction = action('test', ['thisFile']);
+                const result = calculateActionEvents(state, fileAction);
+
+                expect(result.hasUserDefinedEvents).toBe(true);
+
+                expect(result.events).toEqual([openQRCodeScanner(false)]);
             });
         });
     });