+++ CHANGELOG.md
@@ -1,5 +1,28 @@
 # AUX Changelog
 
+## V0.5.2
+
+### Date: 04/26/2019
+
+### Changes:
+
+-   Improvements
+    -   Set builder's default background color to dark gray. Player remains the light blue.
+    -   Changed the `onDragAny/onDropAny` actions to be `onAnyDrag/onAnyDrop`.
+    -   `formula-lib.ts` has changed `isPlayerInContext` export to `player.isInContext`.
+    -   `formula-lib.ts` has changed `makeDiff` export to `diff`.
+    -   Made the mini file dots much smaller.
+    -   Added the ability to show and hide a QR Code Scanner using the `openQRCodeScanner()` and `closeQRCodeScanner()` functions.
+        -   Upon scanning a QR Code the `onQRCodeScanned()` event is triggered with the `that` variable bound to the scanned QR code.
+        -   The `onQRCodeScannerOpened()` event is triggered whenever the QR Code Scanner is opened.
+        -   The `onQRCodeScannerClosed()` event is triggered whenever the QR Code Scanner is closed.
+    -   Moved the file sheet to the right side of the screen.
+-   Bug Fixes
+    -   Fixed an issue with trying to load a WebP version of the "add tag" icon in Safari.
+        -   Safari doesn't support WebP - so we instead have to load it as a PNG.
+    -   Fixed the proxy to return the original content type of images to Safari.
+        -   Because Safari doesn't support WebP we can't automatically optimize the images.
+
 ## V0.5.1
 
 ### Date: 04/25/2019